+++ jgclark.DashboardReact/src/dashboardHelpers.js
@@ -1,7 +1,7 @@
 // @flow
 //-----------------------------------------------------------------------------
 // Dashboard plugin helper functions
-// Last updated 21.6.2024 for v2.0.0-b10 by @jgclark
+// Last updated 28.6.2024 for v2.0.0-b15 by @jgclark
 //-----------------------------------------------------------------------------
 
 import moment from 'moment/min/moment-with-locales'
@@ -9,11 +9,14 @@ import pluginJson from '../plugin.json'
 import { addChecklistToNoteHeading, addTaskToNoteHeading } from '../../jgclark.QuickCapture/src/quickCapture'
 import { allSectionDetails } from "./constants"
 import { parseSettings } from './shared'
-import type { TActionOnReturn, TBridgeClickHandlerResult, TItemType, TParagraphForDashboard, TSection } from './types'
+import type { TActionOnReturn, TBridgeClickHandlerResult, TItemType, TDoneCounts, TParagraphForDashboard, TSection } from './types'
 import { getParaAndAllChildren } from '@helpers/blocks'
 import {
-  removeDateTagsAndToday, getAPIDateStrFromDisplayDateStr, includesScheduledFutureDate, getTodaysDateHyphenated
-  // getISODateStringFromYYYYMMDD
+  getAPIDateStrFromDisplayDateStr,
+  getDateStringFromCalendarFilename,
+  getTodaysDateHyphenated,
+  includesScheduledFutureDate,
+  removeDateTagsAndToday,
 } from '@helpers/dateTime'
 import { clo, JSP, logDebug, logError, logInfo, logWarn, timer } from '@helpers/dev'
 import { createRunPluginCallbackUrl, displayTitle } from '@helpers/general'
@@ -209,6 +212,55 @@ export function makeDashboardParas(origParas: Array<TParagraph>): Array<TParagra
 //-----------------------------------------------------------------
 
 /**
+ * Return number of completed tasks in the note
+ * @param {string} filename
+ * @param {boolean} useEditorWherePossible? use the open Editor to read from if it happens to be open
+ * @returns {[Array<TParagraph>, Array<TParagraph>]} see description above
+ */
+export function getNumCompletedTasksTodayFromNote(filename: string, useEditorWherePossible?: boolean): TDoneCounts {
+  try {
+    let parasToUse: $ReadOnlyArray<TParagraph>
+
+    //------------------------------------------------
+    // Get paras from the note
+    if (useEditorWherePossible && Editor && Editor?.note?.filename === filename) {
+      // If note of interest is open in editor, then use latest version available, as the DataStore could be stale.
+      parasToUse = Editor.paragraphs
+      logDebug('getNumCompletedTasksTodayFromNote', `Using EDITOR (${Editor.filename}) for note '${filename}`)
+    } else {
+      // read note from DataStore in the usual way
+      let note = DataStore.projectNoteByFilename(filename)
+      if (!note) {
+        note = DataStore.calendarNoteByDateString(getDateStringFromCalendarFilename(filename))
+      }
+      if (!note) throw new Error(`Note not found: ${filename}`)
+      parasToUse = note.paragraphs
+      logDebug('getNumCompletedTasksTodayFromNote', `Processing ${note.filename}`)
+    }
+
+    // Calculate the number of closed items
+    const todayHyphenated = getTodaysDateHyphenated()
+    const RE_DONE_TODAY = new RegExp(`@done\\(${todayHyphenated}.*\\)`)
+    const numCompletedTasks = parasToUse.filter((p) => (p.type === 'done') && RE_DONE_TODAY.test(p.content)).length
+
+    const outputObject: TDoneCounts = {
+      completedTasks: numCompletedTasks,
+      // completedChecklists: numCompletedChecklists,
+      lastUpdated: new Date(),
+    }
+    logDebug('getNumCompletedTasksTodayFromNote', `-> ${String(numCompletedTasks)}`)
+    return outputObject
+  } catch (error) {
+    logError('getNumCompletedTasksTodayFromNote', error.message)
+    return {
+      completedTasks: 0,
+      // completedChecklists: 0,
+      lastUpdated: new Date(),
+    }
+  }
+}
+
+/**
  * Return list(s) of open task/checklist paragraphs in calendar note of type 'timePeriodName', or scheduled to that same date.
  * Various config.* items are used:
  * - ignoreFolders? for folders to ignore for referenced notes
@@ -219,7 +271,7 @@ export function makeDashboardParas(origParas: Array<TParagraph>): Array<TParagra
  * @param {string} timePeriodName
  * @param {TNote} timePeriodNote base calendar note to process
  * @param {dashboardConfigType} config
- * @param {boolean} useEditorWherePossible? use the open Editor to read from it happens to be open
+ * @param {boolean} useEditorWherePossible? use the open Editor to read from if it happens to be open
  * @returns {[Array<TParagraph>, Array<TParagraph>]} see description above
  */
 export function getOpenItemParasForCurrentTimePeriod(
@@ -337,8 +389,6 @@ export function getOpenItemParasForCurrentTimePeriod(
 
     // Sort the list by priority then time block, otherwise leaving order the same
     // Then decide whether to return two separate arrays, or one combined one
-    // Note: This takes 100ms
-    // TODO: extend to deal with 12hr (AM/PM) time blocks
     if (config.separateSectionForReferencedNotes) {
       const sortedOpenParas = sortListBy(openDashboardParas, ['-priority', 'timeStr'])
       const sortedRefOpenParas = sortListBy(refOpenDashboardParas, ['-priority', 'timeStr'])
@@ -443,44 +493,6 @@ function isTimeBlockPara(para: TParagraph, mustContainStringArg: string = ''): b
   return !isTermInURL(tbString, para.content)
 }
 
-/** 
- * Display time blocks with .timeBlock style
-* Note: uses definition of time block syntax from plugin helpers, not directly from NP itself. So it may vary slightly.
-* Note: copy from HTMLView.js to avoid React problem
-* @param {string} input
-* @returns {string}
-* FIXME: why is this not used?
- */
-function convertTimeBlockToHTML(input: string): string {
-  let output = input
-  if (isTimeBlockLine(input)) {
-    const timeBlockPart = getTimeBlockString(input)
-    logDebug(`found time block '${timeBlockPart}'`)
-    output = output.replace(timeBlockPart, `<span class="timeBlock">${timeBlockPart}</span>`)
-  }
-  return output
-}
-
-// /**
-//  * Parses and sorts dates from items based on the content field.
-//  * @author @jgclark, @dwertheimer, ChatGPT
-//  * @param {Array<TParagraph>} items - Array of Paragraphs with a content field.
-//  * @returns {Array<TParagraph>} - Array of Paragraphs sorted by the computed start time represented in the text, ignoring ones that do not contain times.
-//  */
-// function parseAndSortDates(items: Array<TParagraph>): Array<ParsedTextDateRange> {
-//   const withDates = items
-//     .map((item) => ({
-//       item,
-//       date: Calendar.parseDateText(item.content)[0]?.start ?? null,
-//     })) // Map each item to an object including both the item and the parsed start date.
-//     .filter(({ date }) => date != null) // Filter out items without a valid start date.
-
-//   // Sort the intermediate structure by the start date and map back to the original items.
-//   const sortedItems = withDates.sort((a, b) => a.date - b.date).map(({ item }) => item)
-
-//   return sortedItems
-// }
-
 /**
  * @params {dashboardConfigType} config Settings
  * @returns {}
@@ -563,7 +575,7 @@ export function makeNoteTitleWithOpenActionFromNPDateStr(NPDateStr: string, item
 }
 
 /**
- * FIXME: write some tests
+ * TODO: write some tests
  * Extend the paragraph objects with a .timeStr property which comes from the start time of a time block, or else 'none' (which will then sort after times).
  * Copes with 'AM' and 'PM' suffixes. Note: Not fully internationalised (but then I don't think the rest of NP accepts non-Western numerals)
  * @tests in dashboardHelpers.test.js
@@ -659,27 +671,6 @@ export function makeFakeCallbackButton(buttonText: string, pluginName: string, c
 }
 
 /**
- * WARNING: DEPRECATED in favour of newer makePluginCommandButton() in HTMLView.js
- * Make HTML for a real button that is used to call one of this plugin's commands.
- * Note: this is not a real button, bcause at the time I started this real <button> wouldn't work in NP HTML views, and Eduard didn't know why.
- * V2: send params for an invokePluginCommandByName call
- * V1: send URL for x-callback
- * @param {string} buttonText to display on button
- * @param {string} pluginName of command to call
- * @param {string} commandName to call when button is 'clicked'
- * @param {string} commandArgs (may be empty)
- * @param {string?} tooltipText to hover display next to button
- * @returns {string}
- */
-export function makeRealCallbackButton(buttonText: string, pluginName: string, commandName: string, commandArgs: string, tooltipText: string = ''): string {
-  const xcallbackURL = createRunPluginCallbackUrl(pluginName, commandName, commandArgs)
-  const output = tooltipText
-    ? `<button class="XCBButton tooltip"><a href="${xcallbackURL}">${buttonText}</a><span class="tooltiptext">${tooltipText}</span></button>`
-    : `<button class="XCBButton"><a href="${xcallbackURL}">${buttonText}</a></button>`
-  return output
-}
-
-/**
  * Move a task or checklist from one calendar note to another.
  * It's designed to be used when the para itself is not available; the para will try to be identified from its filename and content, and it will throw an error if it fails.
  * If 'headingToPlaceUnder' is provided, para is added after it (with heading being created at effective top of note if necessary).
@@ -711,7 +702,7 @@ export async function moveItemBetweenCalendarNotes(NPFromDateStr: string, NPToDa
     }
     const matchedPara = possiblePara
     const itemType = matchedPara?.type
-    const matchedParaAndChildren = getParaAndAllChildren(matchedPara)
+    const matchedParaAndChildren = getParaAndAllChildren(matchedPara) // FIXME: should this be used?
 
     // Remove any scheduled date on the item
     const targetContent = removeDateTagsAndToday(paraContent, true)
@@ -886,4 +877,29 @@ export function mergeSections(existingSections: Array<TSection>, newSections: Ar
     }
   })
   return existingSections
+}
+
+/**
+ * Updates the total count of completed tasks and checklists by adding the counts from all available sections.
+ * TODO: Extend to cover other notes updated today
+ * @param {Array<TSection>} sections
+ */
+export function getTotalDoneCounts(sections: Array<TSection>): TDoneCounts {
+  let numDoneTasks = 0
+  // let numDoneChecklists = 0
+  let latestDate: Date = new Date(0)
+  for (const thisSection of sections) {
+    const thisDC = thisSection.doneCounts
+    if (thisDC) {
+      numDoneTasks += thisDC.completedTasks
+      // numDoneChecklists += thisDC.completedChecklists
+      if (thisDC.lastUpdated > latestDate) latestDate = thisDC.lastUpdated
+    }
+  }
+  logDebug('getTotalDoneCounts', `-> numDoneTasks = ${numDoneTasks} / latestDate = ${latestDate.toLocaleTimeString()}`)
+  return {
+    completedTasks: numDoneTasks,
+    // completedChecklists: numDoneChecklists,
+    lastUpdated: latestDate
+  }
 }