+++ jgclark.DashboardReact/src/types.js
@@ -1,7 +1,7 @@
 // @flow
 //-----------------------------------------------------------------------------
 // Types for Dashboard code
-// Last updated 20.6.2024 for v2.0.0-b10 by @jgclark
+// Last updated 28.6.2024 for v2.0.0-b15 by @jgclark
 //-----------------------------------------------------------------------------
 
 export type TSectionCode = 'DT' | 'DY' | 'DO' | 'W' | 'M' | 'Q' | 'OVERDUE' | 'TAG' | 'PROJ' // | 'COUNT' // where DT = today, DY = yesterday, TAG = Tag, PROJ = Projects section
@@ -22,6 +22,7 @@ export type TSection = {
   actionButtons?: Array<TActionButton>,
   generatedDate?: Date, // note different from lastFullRefresh on whole project
   totalCount?: number, // for when not all possible items are passed in pluginData
+  doneCounts?: TDoneCounts, // number of tasks and checklists completed today etc.
 }
 
 export type TItemType = 'open' | 'checklist' | 'congrats' | 'project' | 'filterIndicator'
@@ -65,6 +66,12 @@ export type TProjectForDashboard = {
   reviewInterval: string, /* from the Project instance */
   percentComplete: number, /* from the Project instance */
   lastProgressComment: string, /* from the Project instance */
+}
+
+export type TDoneCounts = {
+  completedTasks: number,
+  // completedChecklists: number,
+  lastUpdated: Date
 }
 
 // details for a UI button
@@ -209,8 +216,8 @@ export type TPluginData = {
   lastFullRefresh: Date, /* localized date string new Date().toLocaleString() */
   themeName: string, /* the theme name used when generating the dashboard */
   platform: string, /* the platform used when generating the dashboard */
-  demoMode: boolean, /* generate fake content */
-  doneCount?: number,
+  demoMode: boolean, /* use fake content for demo purposes */
+  totalDoneCounts?: TDoneCounts,
   startDelayedRefreshTimer?: boolean, /* start the delayed refresh timer hack set in post processing commands*/
 }